// frontend/src/app/models/register-request.model.ts
export interface RegisterRequest {
    firstName: string;
    lastName: string;
    email: string;
    username: string;
    password: string;
    birthDate: string;
    role: string;
}

`
{
  "firstName": "Test 3",
  "lastName":  "TestTest 3",
  "email": "tester3@mail.com",
  "username": "tester3",
  "password": "12345678",
  "birthDate": "2003-03-08",
  "role": "ADMIN"
}`